Logger App: Chart printing. Firmware: integer to char array, with neg. numbers.

+++ Statemachine_ST7783_ADC/main.h
@@ -6,5 +6,6 @@
 
 #define CHARTONUM(a, b) ((a - '0') * b)
 
-void bin2asc(uint16_t value, char buffer[], uint8_t digits);
+//void bin2asc(uint16_t value, char buffer[], uint8_t digits);
+void bin2asc(int value, char buffer[], uint8_t digits);
